requirements.txt
# Dependências para o sistema NFSe Arapiraca
zeep==4.2.1
lxml==4.9.3
xmlsec==1.3.13
python-dotenv==1.0.0
cryptography==41.0.7
requests==2.31.0